Extract makeRouteInfo logic
